A predictive model for forecasting demand and supply information of TOP crops
India is majorly an agriculture-based economy. Around 42% of the people depend on agriculture for their
livelihood. The economic upliftment of farmers happens when there is a seamless transfer of agricultural
produce from producers to the consumers. It is evident that there is a huge gap between demand and supply of
various crops, due to which both farmers and consumers are facing problems. At present, in India there is no
system in place to efficiently manage this demand and supply issue. The potential of present-day technologies
like data analytics, machine learning can be exploited to overcome these issues. The available data about the
demand, supply, price variation of the crops and other factors affecting the supply chain of agricultural produce
can be used to analyse and come up with a model to predict and forecast market variations of agricultural
crops. The proposed work intends to analyse the various authentic data available for TOP (Tomato, Potato and
Onion) crops and design a supply-demand prediction model to forecast the market fluctuations as an advisory
measure.